// scripts/action/types.ts
export type HooksRegistration = {
  beforeInitialize?: boolean
  afterInitialize?: boolean
  beforeAddLiquidity?: boolean
  afterAddLiquidity?: boolean
  beforeRemoveLiquidity?: boolean
  afterRemoveLiquidity?: boolean
  beforeSwap?: boolean
  afterSwap?: boolean
  beforeDonate?: boolean
  afterDonate?: boolean
  beforeSwapReturnsDelta?: boolean
  afterSwapReturnsDelta?: boolean
  afterMintReturnsDelta?: boolean
  afterBurnReturnsDelta?: boolean
}

export enum POOL_TYPE {
  CLAMM = 'CL',
  Bin = 'Bin',
}
// export type PoolType = 'CL' | 'Bin'
export type PoolType = `${POOL_TYPE}`

export type CLPoolParameter = {
  /**
   * Hooks registration for the pool
   * @see {@link HooksRegistration}
   */
  hooksRegistration?: HooksRegistration
  tickSpacing: number
}

export type BinPoolParameter = {
  /**
   * Hooks registration for the pool
   * @see {@link HooksRegistration}
   */
  hooksRegistration?: HooksRegistration
  binStep: number
}

/**
 * PoolKey is a unique identifier for a pool
 *
 * decoded version of `PoolKey`
 *
 */
export type PoolKey = {
  /**
   * the lower currency address of the pool, use zero address for native token
   */
  currency0: Address
  /**
   * the higher currency address of the pool
   */
  currency1: Address
  /**
   * the address of the hooks contract, if not set, use zero address
   */
  hooks?: Address
  /**
   * the lp fee of the pool, the max fee for cl pool is 1_000_000(100%) and for bin, it is 100_000(10%).
   * If the pool has dynamic fee then it must be exactly equal to 0x800000
   *
   * @see DYNAMIC_FEE_FLAG
   */
  fee: number
  /**
   * the parameters of the pool
   * include:
   *   1. hooks registration callback
   *   2. pool specific parameters: tickSpacing for CLPool, binStep for BinPool
   *
   * @see BinPoolParameter
   * @see CLPoolParameter
   * @see HooksRegistration
   */
  parameters: CLPoolParameter
}